Module epoll::util::ctl_op

Constants

ADD
Register the target file descriptor fd on the epoll instance referred to by the file descriptor epfd and associate the event with the internal file linked to fd.

DEL
Remove (deregister) the target file descriptor fd from the epoll instance referred to by epfd. The event is ignored and can be NULL.

MOD
Change the event event associated with the target file descriptor fd.