Enum epoll::util::CtlError

pub enum CtlError {
    EBADF,
    EEXIST,
    EINVAL,
    ENOENT,
    ENOSPC,
    EPERM,
}

Variants

EBADF

fd is not a valid file descriptor.

EEXIST

op was EPOLL_CTL_ADD, and the supplied file descriptor fd is already registered with this epoll instance.

EINVAL

epfd is not an epoll file descriptor, or fd is the same as epfd, or the requested operation op is not supported by this interface.

ENOENT

op was EPOLL_CTL_MOD or EPOLL_CTL_DEL, and fd is not registered with this epoll instance.

ENOSPC

There was insufficient memory to handle the requested op control operation.

EPERM

The limit imposed by /proc/sys/fs/epoll/max_user_watches was encountered while trying to register (EPOLL_CTL_ADD) a new file descriptor on an epoll instance.