Enum epoll::util::WaitError

pub enum WaitError {
    EBADF,
    EFAULT,
    EINTR,
    EINVAL,
}

Variants

EBADF

epfd is not a valid file descriptor.

EFAULT

The memory area pointed to by events is not accessible with write permissions.

EINTR

The call was interrupted by a signal handler before either any of the requested events occurred or the timeout expired.

EINVAL

epfd is not an epoll file descriptor, or maxevents is less than or equal to zero.